move EnableStartPage enum to groups.ts
