+++ src/client/datascience/commands/commandRegistry.ts
@@ -8,7 +8,8 @@ import { CodeLens, ConfigurationTarget, env, Range, Uri } from 'vscode';
 import { ICommandNameArgumentTypeMapping } from '../../common/application/commands';
 import { IApplicationShell, ICommandManager, IDebugService, IDocumentManager } from '../../common/application/types';
 import { Commands as coreCommands } from '../../common/constants';
-import { EnableStartPage, IStartPage } from '../../common/startPage/types';
+import { EnableStartPage } from '../../common/experiments/groups';
+import { IStartPage } from '../../common/startPage/types';
 import { IConfigurationService, IDisposable, IExperimentService, IOutputChannel } from '../../common/types';
 import { DataScience } from '../../common/utils/localize';
 import { noop } from '../../common/utils/misc';
